Add TestParkingSpotTypeService.java

+++ src/main/java/ca/mcgill/ecse321/parkinglotsystem/service/ParkingSpotTypeService.java
@@ -1,22 +1,20 @@
 package ca.mcgill.ecse321.parkinglotsystem.service;
 
-import java.sql.Date;
-import java.sql.Time;
+
 import java.util.ArrayList;
 import java.util.List;
-
+import org.springframework.http.HttpStatus;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jmx.access.InvalidInvocationException;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import ca.mcgill.ecse321.parkinglotsystem.dao.ParkingSpotRepository;
 import ca.mcgill.ecse321.parkinglotsystem.dao.ParkingSpotTypeRepository;
-import ca.mcgill.ecse321.parkinglotsystem.dto.ParkingSpotTypeDto;
 import ca.mcgill.ecse321.parkinglotsystem.model.ParkingSpot;
 import ca.mcgill.ecse321.parkinglotsystem.model.ParkingSpotType;
 import ca.mcgill.ecse321.parkinglotsystem.service.utilities.HelperMethods;
-
+import ca.mcgill.ecse321.parkinglotsystem.service.exceptions.CustomException;
 
 @Service
 public class ParkingSpotTypeService {
@@ -44,7 +42,7 @@ public class ParkingSpotTypeService {
         ParkingSpotType parkingSpotType = new ParkingSpotType();
 
         if (error.length() > 0) {
-			throw new IllegalArgumentException(error);
+			throw new CustomException(error, HttpStatus.BAD_REQUEST);
 		}
 
         else {
@@ -70,11 +68,10 @@ public class ParkingSpotTypeService {
 
         // Input validation
         String error = "";
+        boolean canDelete = true;
         if (name == null || name.trim().length() == 0) {
             error = error + "a name must be mention to delete parking spot type! ";
         }
-
-
         ParkingSpotType parkingSpotType = parkingSpotTypeRepository.findParkingSpotTypeByName(name);
 
         if (parkingSpotType == null) {
@@ -82,20 +79,30 @@ public class ParkingSpotTypeService {
         }
 
         if (error.length() >0) {
-			throw new IllegalArgumentException(error);
+			throw new CustomException(error, HttpStatus.BAD_REQUEST);
 		}
 
         //we must delete the parking spot as a parking spot must have a parking spot type
-        if (parkingSpotRepository.findAll() != null) {
-            Iterable<ParkingSpot> parkingSpots = parkingSpotRepository.findAll();
-            for (ParkingSpot p: parkingSpots) {
-                if (p.getType().getName() == name) {
-                    parkingSpotRepository.delete(p);
-                }
-            }
-
+        List<ParkingSpot> parkingSpots = new ArrayList<ParkingSpot>();
+        parkingSpots = parkingSpotRepository.findParkingSpotByType(parkingSpotType);
+        if (parkingSpots.size() >0 ) {
+            canDelete = false;
+            throw new CustomException("This type is assigned to a parking spot ", HttpStatus.BAD_REQUEST);
         }
-        parkingSpotTypeRepository.delete(parkingSpotType);
+        // if (parkingSpotRepository.findAll() != null) {
+        //     Iterable<ParkingSpot> parkingSpots = parkingSpotRepository.findAll();
+        //     for (ParkingSpot p: parkingSpots) {
+        //         if (p.getType().getName() == name) {
+        //             canDelete = false;
+        //             throw new CustomException("This type is assigned to a parking spot ", HttpStatus.BAD_REQUEST);
+        //         }
+        //     }
+
+        // }
+        if (canDelete) {
+            parkingSpotTypeRepository.delete(parkingSpotType);
+        }
+        
 
         return parkingSpotType;
     }
@@ -120,7 +127,7 @@ public class ParkingSpotTypeService {
             error = error + "Could not find a parking spot type by this name to update! ";
         }
         if (error.length()>0) {
-			throw new IllegalArgumentException(error);
+			throw new CustomException(error, HttpStatus.BAD_REQUEST);
 		}
 
         parkingSpotType.setFee(fee);
@@ -138,7 +145,7 @@ public class ParkingSpotTypeService {
         }
 
         if (error.length()>0) {
-			throw new IllegalArgumentException(error);
+			throw new CustomException(error, HttpStatus.BAD_REQUEST);
 		}
 
         ParkingSpotType parkingSpotType = parkingSpotTypeRepository.findParkingSpotTypeByName(name);
